Revert "chore(refactor): Refactor router callbacks into EntryControllermethods"
This reverts commit b7809325624d94470ba03a2e4f1a7510d5419a28.
